+++ src/compiler/checker/lines_24000_32000.rs
@@ -1,0 +1,262 @@
+#![allow(non_upper_case_globals)]
+
+use bitflags::bitflags;
+use std::borrow::Borrow;
+use std::cell::{Ref, RefCell, RefMut};
+use std::collections::{HashMap, HashSet};
+use std::ptr;
+use std::rc::Rc;
+
+use crate::{
+    TypeNode, UnionOrIntersectionType, UnionOrIntersectionTypeInterface, UnionType,
+    VariableDeclaration, VariableStatement, __String, bind_source_file, chain_diagnostic_messages,
+    create_diagnostic_collection, create_diagnostic_for_node,
+    create_diagnostic_for_node_from_message_chain, create_printer, create_symbol_table,
+    create_text_writer, declaration_name_to_string, escape_leading_underscores, every, factory,
+    first_defined, first_or_undefined, for_each, get_effective_initializer,
+    get_effective_type_annotation_node, get_first_identifier, get_name_of_declaration,
+    get_object_flags, get_source_file_of_node, get_synthetic_factory, has_dynamic_name,
+    has_initializer, is_binding_element, is_external_or_common_js_module, is_identifier_text,
+    is_object_literal_expression, is_private_identifier, is_property_assignment,
+    is_property_declaration, is_property_signature, is_variable_declaration, node_is_missing,
+    object_allocator, unescape_leading_underscores, using_single_line_string_writer, ArrayTypeNode,
+    BaseInterfaceType, BaseIntrinsicType, BaseLiteralType, BaseNodeFactorySynthetic,
+    BaseObjectType, BaseType, BaseUnionOrIntersectionType, CharacterCodes, Debug_, Diagnostic,
+    DiagnosticCollection, DiagnosticMessage, DiagnosticMessageChain, Diagnostics, EmitHint,
+    EmitTextWriter, Expression, ExpressionStatement, FreshableIntrinsicType,
+    HasExpressionInitializerInterface, InterfaceDeclaration, InterfaceType, IntrinsicType,
+    KeywordTypeNode, LiteralLikeNode, LiteralLikeNodeInterface, LiteralType, LiteralTypeInterface,
+    NamedDeclarationInterface, Node, NodeInterface, Number, NumberLiteralType, NumericLiteral,
+    ObjectFlags, ObjectFlagsTypeInterface, ObjectLiteralExpression, PrefixUnaryExpression,
+    PrinterOptions, PropertyAssignment, PropertySignature, RelationComparisonResult,
+    ResolvableTypeInterface, ResolvedTypeInterface, SourceFile, Statement, StringLiteralType,
+    Symbol, SymbolFlags, SymbolFormatFlags, SymbolTable, SymbolTracker, SyntaxKind, Ternary, Type,
+    TypeChecker, TypeCheckerHost, TypeElement, TypeFlags, TypeInterface, TypeReferenceNode,
+    UnionReduction, VariableLikeDeclarationInterface,
+};
+
+impl TypeChecker {
+    pub(crate) fn get_contextual_type_for_variable_like_declaration(
+        &self,
+        declaration: &Node,
+    ) -> Option<Rc<Type>> {
+        let type_node = get_effective_type_annotation_node(declaration);
+        if let Some(type_node) = type_node {
+            return Some(self.get_type_from_type_node(&*type_node));
+        }
+        match declaration.kind() {
+            _ => None,
+        }
+    }
+
+    pub(crate) fn get_contextual_type_for_initializer_expression<TNode: NodeInterface>(
+        &self,
+        node: &TNode,
+    ) -> Option<Rc<Type>> {
+        let parent = node.parent();
+        let declaration = match &*parent {
+            Node::VariableDeclaration(variable_declaration) => variable_declaration,
+            _ => panic!("Expected VariableDeclaration"),
+        };
+        if has_initializer(declaration)
+            && Rc::ptr_eq(&node.node_wrapper(), &declaration.initializer().unwrap())
+        {
+            let result = self.get_contextual_type_for_variable_like_declaration(&*parent);
+            if result.is_some() {
+                return result;
+            }
+        }
+        None
+    }
+
+    pub(crate) fn get_type_of_property_of_contextual_type(
+        &self,
+        type_: Rc<Type>,
+        name: &__String,
+    ) -> Option<Rc<Type>> {
+        self.map_type(
+            type_,
+            &mut |t| {
+                if false {
+                    unimplemented!()
+                } else if t.flags().intersects(TypeFlags::StructuredType) {
+                    let prop = self.get_property_of_type(t, name);
+                    if let Some(prop) = prop {
+                        return if false {
+                            None
+                        } else {
+                            Some(self.get_type_of_symbol(&*prop))
+                        };
+                    }
+                    return if let Some(found) = Option::<()>::None /*self.find_applicable_index_info(self.get_index_infos_of_structured_type(t), self.get_string_literal_type(unescape_leading_underscores(name)))*/ {
+                        unimplemented!()
+                    } else {
+                        None
+                    };
+                }
+                None
+            },
+            Some(true),
+        )
+    }
+
+    pub(crate) fn get_contextual_type_for_object_literal_element(
+        &self,
+        element: &PropertyAssignment,
+    ) -> Option<Rc<Type>> {
+        let parent = element.parent();
+        let object_literal = match &*parent {
+            Node::Expression(Expression::ObjectLiteralExpression(object_literal_expression)) => {
+                object_literal_expression
+            }
+            _ => panic!("Expected ObjectLiteralExpression"),
+        };
+        // let property_assignment_type = if is_property_assignment(element) {
+        // } else {
+        //     None
+        // };
+        // if property_assignment_type.is_some() {
+        //     return property_assignment_type;
+        // }
+        let type_ = self.get_apparent_type_of_contextual_type(object_literal);
+        if let Some(type_) = type_ {
+            if self.has_bindable_name(element) {
+                return self.get_type_of_property_of_contextual_type(
+                    type_,
+                    &self.get_symbol_of_node(element).unwrap().escaped_name,
+                );
+            }
+            unimplemented!()
+        }
+        None
+    }
+
+    pub(crate) fn get_apparent_type_of_contextual_type<TNode: NodeInterface>(
+        &self,
+        node: &TNode, /*Expression | MethodDeclaration*/
+    ) -> Option<Rc<Type>> {
+        let contextual_type = if false {
+            unimplemented!()
+        } else {
+            self.get_contextual_type(node)
+        };
+        let instantiated_type = self.instantiate_contextual_type(contextual_type, node);
+        if let Some(instantiated_type) = instantiated_type {
+            if true {
+                let apparent_type = self
+                    .map_type(
+                        instantiated_type,
+                        &mut |type_| Some(self.get_apparent_type(type_)),
+                        Some(true),
+                    )
+                    .unwrap();
+                return if apparent_type.flags().intersects(TypeFlags::Union)
+                    && is_object_literal_expression(node)
+                {
+                    unimplemented!()
+                } else if false {
+                    unimplemented!()
+                } else {
+                    Some(apparent_type)
+                };
+            }
+        }
+        None
+    }
+
+    pub(crate) fn instantiate_contextual_type<TNode: NodeInterface>(
+        &self,
+        contextual_type: Option<Rc<Type>>,
+        node: &TNode,
+    ) -> Option<Rc<Type>> {
+        if false {
+            unimplemented!()
+        }
+        contextual_type
+    }
+
+    pub(crate) fn get_contextual_type<TNode: NodeInterface>(
+        &self,
+        node: &TNode, /*Expression*/
+    ) -> Option<Rc<Type>> {
+        let parent = node.parent();
+        match &*parent {
+            Node::VariableDeclaration(variable_declaration) => {
+                self.get_contextual_type_for_initializer_expression(node)
+            }
+            Node::PropertyAssignment(property_assignment) => {
+                self.get_contextual_type_for_object_literal_element(property_assignment)
+            }
+            _ => unimplemented!(),
+        }
+    }
+
+    pub(crate) fn check_object_literal(&self, node: &ObjectLiteralExpression) -> Rc<Type> {
+        let mut properties_table = create_symbol_table();
+        let mut properties_array: Vec<Rc<Symbol>> = vec![];
+
+        let object_flags = self.fresh_object_literal_flag;
+
+        for member_decl in &node.properties {
+            let member = self.get_symbol_of_node(&**member_decl).unwrap();
+            if member_decl.kind() == SyntaxKind::PropertyAssignment {
+            } else {
+                unimplemented!()
+            }
+
+            if false {
+                unimplemented!()
+            } else {
+                properties_table.insert(member.escaped_name.clone(), member.clone());
+            }
+            properties_array.push(member);
+        }
+
+        let create_object_literal_type = || {
+            let mut result =
+                self.create_anonymous_type(node.symbol(), Rc::new(RefCell::new(properties_table)));
+            result.set_object_flags(
+                result.object_flags()
+                    | object_flags
+                    | ObjectFlags::ObjectLiteral
+                    | ObjectFlags::ContainsObjectOrArrayLiteral,
+            );
+            Rc::new(result.into())
+        };
+
+        create_object_literal_type()
+    }
+
+    pub(crate) fn is_known_property(
+        &self,
+        target_type: Rc<Type>,
+        name: &__String,
+        is_comparing_jsx_attributes: bool,
+    ) -> bool {
+        if target_type.flags().intersects(TypeFlags::Object) {
+            if self
+                .get_property_of_object_type(target_type, name)
+                .is_some()
+                || false
+            {
+                return true;
+            }
+        } else if target_type
+            .flags()
+            .intersects(TypeFlags::UnionOrIntersection)
+            && self.is_excess_property_check_target(target_type)
+        {
+            unimplemented!()
+        }
+        false
+    }
+
+    pub(crate) fn is_excess_property_check_target(&self, type_: Rc<Type>) -> bool {
+        (type_.flags().intersects(TypeFlags::Object)
+            && !(get_object_flags(&*type_)
+                .intersects(ObjectFlags::ObjectLiteralPatternWithComputedProperties)))
+            || type_.flags().intersects(TypeFlags::NonPrimitive)
+            || (type_.flags().intersects(TypeFlags::Union) && unimplemented!())
+            || (type_.flags().intersects(TypeFlags::Intersection) && unimplemented!())
+    }
+}
